serialize settings on close
Previously, the nodes were serialized on demand. This caused intent preferences to call the target package supplier. The supplier invoked ChangePackageName#setOrGetPackageName and as per design, once it is called, the package name that was set will be preferred over later calls. Moving the serialization to close, meant that any patch that called ChangePackageName#setOrGetPackageName would be preferred to set chose the preferred package name so that on close, the serializer will use the preferred package name of the patches that were executed before.