Clarify reddit Spoof client app instructions
With the kebab menu on the side asking for a file/folder. It can be easily misunderstood by users that this is a file selection prompt. When it actually is a string field asking for the client ID.

This can be even more confusing to users who have their client id inside the ``reddit_client_id_revanced.txt`` file from previous version of this patch.

Preferably, this would be fixed in the app by removing the Kebab menu or making it input the file contents instead. But I am not an android dev.